Measure + test + some bugfixes in tests

+++ src/main/java/org/siery/music/note/ChordNote.java
@@ -4,20 +4,20 @@ import org.siery.music.chord.Chord;
 
 public class ChordNote extends Chord implements Note {
 
-	protected float duration;
+	protected double duration;
 	
-	public ChordNote(Chord chord, float duration) {
+	public ChordNote(Chord chord, double duration) {
 		this.duration = duration;
 		this.tones = chord.getTones();
 	}
 	
 	@Override
-	public float getDuration() {
+	public double getDuration() {
 		return duration;
 	}
 
 	@Override
-	public void setDuration(float duration) {
+	public void setDuration(double duration) {
 		this.duration = duration;
 	}
