+++ src/main/java/org/siery/music/note/Note.java
@@ -1,6 +1,6 @@
 package org.siery.music.note;
 
 public interface Note {
-	public void setDuration(float duration);
-	public float getDuration();
+	public void setDuration(double duration);
+	public double getDuration();
 }
